Spinner Project
This is in a zip file because the file has more than 100. This project is to make a list inside a spinner then every time the user clicks to add or constantly clicks on, the list will be printed on a text view where you can add more to it even if it rotated.